✨ 🐛 Fix search filter bug in ProductsComponent
This commit fixes a bug in the `updateFilter` method of the `ProductsComponent`
where the search filter was not being updated correctly. The bug was causing the
search filter to not be set to an empty string when the filter length was less
than or equal to 3 characters.

To fix this, the code now checks the length of the filter and assigns an empty
string if the length is less than or equal to 3, otherwise it assigns the filter
value.

Changes:
- src/app/pages/products/products.component.ts
  - Line 52: Replace the if-else statement with a ternary operator to assign the
    correct value to `filterValue`.

- src/app/pages/products/products.component.scss
  - Line 7: Change the background color of the search component to
    `lightslategray`.

- src/app/pages/products/products.component.html
  - Line 2: Update the search